geometry/pyproject.toml
[project]
name = "geometry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "h5py>=3.16.0",
    "numpy",
    "osmnx",
    "scipy>=1.17.1",
    "triangle>=20250106",
    "trimesh",
    "overturemaps>=0.8.0",
    "pillow>=10.0",
]
